Mask setup generated secrets
