+++ server/route/v2/admin.ts
@@ -1,107 +1,34 @@
 import { Hono } from 'hono';
-import mainJson from '../../json/main.json';
 import {
   authenticateJWT,
   optionalJWT,
   requireAdmin,
   requireSuperAdmin,
 } from '../../middleware/jwtAuth';
-import type {
-  AiConfig,
-  ConfigTestResult,
-  InitializationStatus,
-  SetupRequest,
-  SetupResponse,
-} from '../../types/config';
+import type { AiConfig, ConfigTestResult } from '../../types/config';
 import type { HonoContext, HonoVariables } from '../../types/hono';
-import { UserRole } from '../../types/main';
 import {
   generateSecurityKeys,
   getAllConfigs,
   getConfig,
-  getMissingRequiredConfigs,
   isInitialized,
   refreshConfigCache,
   setConfig,
 } from '../../utils/config';
 import { resolveIncomingAiConfig, sanitizeAiConfig } from '../../utils/ai/providers';
 import { ConfigTester } from '../../utils/configTester';
-import {
-  checkDatabaseConnection,
-  createAdminUser,
-  deleteUserById,
-  findUserByUsernameOrEmail,
-  getLatestMigrationVersion,
-  getRoleStats,
-  getUserByIdForAdmin,
-  hasAdminUser,
-  listUsers,
-  unverifyUserEmail,
-  updateUserRole,
-  verifyUserEmail,
-  getDashboardStats,
-} from '../../utils/dbMethods';
 import { createResponse, getApiUrl } from '../../utils/main';
+import {
+  getStorageFriendlyError,
+  logStorageTestResult,
+  logStorageTestStart,
+} from './adminStorageUtils';
+import adminSetupRouter from './adminSetup';
+import adminUsersRouter from './adminUsers';
 
 const adminRouter = new Hono<{ Variables: HonoVariables }>();
 
-const redactStorageConfig = (config?: any) => ({
-  endpoint: config?.endpoint,
-  bucket: config?.bucket,
-  region: config?.region,
-  urlPrefix: config?.urlPrefix,
-  hasAccessKeyId: Boolean(config?.accessKeyId),
-  hasSecretAccessKey: Boolean(config?.secretAccessKey),
-});
-
-const getStorageFriendlyError = (details: any, fallbackMessage?: string) => {
-  const code = details?.name || details?.Code || details?.code;
-  const httpStatus = details?.$metadata?.httpStatusCode;
-  const rawMessage = typeof details?.message === 'string' ? details.message : '';
-
-  if (code === 'NoSuchBucket' || httpStatus === 404) {
-    return 'Bucket not found. Please confirm the bucket name and region.';
-  }
-
-  if (code === 'AccessDenied' || httpStatus === 403) {
-    return 'Access denied. Please check access keys and bucket permissions.';
-  }
-
-  if (code === 'InvalidAccessKeyId' || code === 'SignatureDoesNotMatch') {
-    return 'Invalid credentials. Please verify access key and secret.';
-  }
-
-  if (code === 'PermanentRedirect' || code === 'MovedPermanently') {
-    return 'Endpoint or region mismatch. Please verify the endpoint and region.';
-  }
-
-  if (
-    code === 'UnknownEndpoint' ||
-    rawMessage.includes('Inaccessible host') ||
-    rawMessage.includes('ENOTFOUND')
-  ) {
-    return 'Unable to reach the endpoint. Please verify the endpoint address and network.';
-  }
-
-  if (
-    code === 'NetworkingError' ||
-    rawMessage.includes('ECONNREFUSED') ||
-    rawMessage.includes('ETIMEDOUT') ||
-    rawMessage.includes('EHOSTUNREACH')
-  ) {
-    return 'Network error while connecting to the endpoint. Please check connectivity.';
-  }
-
-  if (rawMessage) {
-    return rawMessage;
-  }
-
-  if (fallbackMessage) {
-    return fallbackMessage;
-  }
-
-  return 'Please verify the endpoint, bucket, and credentials.';
-};
+adminRouter.route('/', adminSetupRouter);
 
 const sanitizeSettingsForAdmin = (configs: Record<string, any>) => {
   if (!configs.ai) return configs;
@@ -110,306 +37,6 @@ const sanitizeSettingsForAdmin = (configs: Record<string, any>) => {
     ai: sanitizeAiConfig(configs.ai),
   };
 };
-
-const logStorageTestStart = (source: string, config: any) => {
-  console.info('[storage-test] start', source, redactStorageConfig(config));
-};
-
-const logStorageTestResult = (source: string, result: ConfigTestResult) => {
-  if (result.success) {
-    console.info('[storage-test] success', source, result.details || {});
-    return;
-  }
-
-  const metadata = (result.details as any)?.$metadata;
-  console.warn('[storage-test] failed', source, {
-    message: result.message,
-    httpStatusCode: metadata?.httpStatusCode,
-    requestId: metadata?.requestId,
-    extendedRequestId: metadata?.extendedRequestId,
-    details: result.details,
-  });
-};
-
-// 获取系统初始化状态
-adminRouter.get('/status', async (c: HonoContext) => {
-  try {
-    // 检查系统是否已初始化
-    const initialized = await isInitialized();
-
-    // 获取缺失的必需配置
-    const missingRequiredConfigs = await getMissingRequiredConfigs();
-
-    // 检查数据库连接
-    const databaseConnected = await checkDatabaseConnection();
-
-    // 检查是否有管理员用户
-    const hasAdmin = await hasAdminUser();
-
-    // 生成警告信息
-    const warnings: string[] = [];
-
-    if (!databaseConnected) {
-      warnings.push('Database connection failed, please check database configuration');
-    }
-
-    if (initialized && !hasAdmin) {
-      warnings.push('System is initialized but no admin user found');
-    }
-
-    if (missingRequiredConfigs.length > 0) {
-      warnings.push(`Missing required configurations: ${missingRequiredConfigs.join(', ')}`);
-    }
-
-    // 构建响应数据
-    const status: InitializationStatus = {
-      isInitialized: initialized,
-      missingRequiredConfigs,
-      warnings,
-    };
-
-    // 如果系统已初始化，添加额外信息
-    if (initialized) {
-      const systemConfig = await getConfig('system');
-      const siteConfig = await getConfig('site');
-
-      return c.json(
-        createResponse({
-          ...status,
-          systemInfo: {
-            databaseConnected,
-            hasAdmin,
-            siteName: (siteConfig as any)?.name || 'Not set',
-            frontendUrl: (siteConfig as any)?.frontendUrl || 'Not set',
-            initializationVersion: (systemConfig as any)?.initializationVersion || '1.0.0',
-          },
-        }),
-        200
-      );
-    } else {
-      return c.json(createResponse(status), 200);
-    }
-  } catch (error) {
-    console.error('获取系统状态失败:', error);
-    return c.json(createResponse(null, 'Failed to get system status'), 500);
-  }
-});
-
-// 系统初始化向导
-adminRouter.post('/setup', async (c: HonoContext) => {
-  try {
-    // 检查系统是否已经初始化
-    const alreadyInitialized = await isInitialized();
-    if (alreadyInitialized) {
-      return c.json(createResponse(null, 'System has already been initialized'), 400);
-    }
-
-    // 验证请求数据
-    const body = await c.req.json();
-    const setupData: SetupRequest = body;
-    if (!setupData.site || !setupData.ui || !setupData.admin) {
-      return c.json(createResponse(null, 'Missing required configuration data'), 400);
-    }
-
-    // 验证必需字段
-    if (!setupData.site.name || !setupData.site.frontendUrl) {
-      return c.json(createResponse(null, 'Site name and frontend URL are required'), 400);
-    }
-
-    // 存储配置是可选的，但如果提供了，则必须完整
-    if (setupData.storage) {
-      if (
-        !setupData.storage.endpoint ||
-        !setupData.storage.bucket ||
-        !setupData.storage.accessKeyId ||
-        !setupData.storage.secretAccessKey ||
-        !setupData.storage.urlPrefix
-      ) {
-        return c.json(createResponse(null, 'Storage configuration is incomplete'), 400);
-      }
-
-      // 测试存储配置
-      logStorageTestStart('setup', setupData.storage);
-      const storageTest = await ConfigTester.testStorage(setupData.storage);
-      logStorageTestResult('setup', storageTest);
-      if (!storageTest.success) {
-        const friendlyMessage = getStorageFriendlyError(storageTest.details, storageTest.message);
-        return c.json(
-          createResponse(null, `Storage configuration test failed: ${friendlyMessage}`),
-          400
-        );
-      }
-    }
-
-    // 验证 UI 配置
-    if (
-      setupData.ui.allowRegistration === undefined ||
-      setupData.ui.allowUploadFile === undefined ||
-      !setupData.ui.defaultUserRole ||
-      setupData.ui.apiRateLimit === undefined ||
-      setupData.ui.maxVideoUploadSizeMB === undefined
-    ) {
-      return c.json(createResponse(null, 'UI configuration is incomplete'), 400);
-    }
-    // 验证 apiRateLimit 必须 >= 10
-    if (typeof setupData.ui.apiRateLimit !== 'number' || setupData.ui.apiRateLimit < 10) {
-      return c.json(createResponse(null, 'API rate limit must be a number and at least 10'), 400);
-    }
-    if (
-      typeof setupData.ui.maxVideoUploadSizeMB !== 'number' ||
-      setupData.ui.maxVideoUploadSizeMB < 1
-    ) {
-      return c.json(
-        createResponse(null, 'Video upload size limit must be a number and at least 1 MB'),
-        400
-      );
-    }
-    // 验证 defaultUserRole 必须是有效角色
-    if (!['user', 'moderator'].includes(setupData.ui.defaultUserRole)) {
-      return c.json(
-        createResponse(null, 'Default user role must be either "user" or "moderator"'),
-        400
-      );
-    }
-
-    if (!setupData.admin.username || !setupData.admin.email || !setupData.admin.password) {
-      return c.json(createResponse(null, 'Admin user information is incomplete'), 400);
-    }
-
-    // 验证管理员密码长度（至少6个字符）
-    if (setupData.admin.password.length < 6) {
-      return c.json(createResponse(null, 'Password must be at least 6 characters'), 400);
-    }
-    if (setupData.admin.password.length > 128) {
-      return c.json(createResponse(null, 'Password cannot exceed 128 characters'), 400);
-    }
-
-    // 验证管理员用户名不能是受保护的路由名称（不区分大小写）
-    const { safeRoutes } = mainJson;
-    if (safeRoutes.includes(setupData.admin.username.toLowerCase())) {
-      return c.json(
-        createResponse(
-          null,
-          'This username is reserved and cannot be used. Please choose another username.'
-        ),
-        400
-      );
-    }
-
-    // 检查管理员用户是否已存在。旧版/半初始化数据库可能已有管理员但缺少 system
-    // settings，此时复用现有管理员完成配置初始化，避免卡在重复用户名。
-    const hasExistingAdmin = await hasAdminUser();
-    const existingUser = await findUserByUsernameOrEmail({
-      username: setupData.admin.username,
-      email: setupData.admin.email,
-    });
-
-    let adminUser: {
-      id: string;
-      username: string;
-      email: string;
-      role: string;
-    };
-    let reusedExistingAdmin = false;
-
-    if (existingUser) {
-      const isExistingAdmin = ['admin', 'super_admin'].includes(existingUser.role);
-      const isSameAdminIdentity =
-        existingUser.username === setupData.admin.username &&
-        existingUser.email === setupData.admin.email;
-
-      if (!isExistingAdmin || !isSameAdminIdentity) {
-        return c.json(createResponse(null, 'Username or email already exists'), 400);
-      }
-
-      adminUser = existingUser;
-      reusedExistingAdmin = true;
-    } else if (hasExistingAdmin) {
-      return c.json(
-        createResponse(
-          null,
-          'An admin user already exists. Please use the existing admin username and email to finish setup.'
-        ),
-        400
-      );
-    } else {
-      // 创建管理员用户（在事务中）
-      adminUser = await createAdminUser({
-        username: setupData.admin.username,
-        email: setupData.admin.email,
-        password: setupData.admin.password,
-        nickname: setupData.admin.nickname,
-      });
-    }
-
-    // 开始初始化流程
-    // 1. 保存站点配置
-    await setConfig(
-      'site',
-      {
-        name: setupData.site.name,
-        frontendUrl: setupData.site.frontendUrl,
-        description: setupData.site.description || '',
-        defaultLanguage: setupData.site.defaultLanguage || 'zh-CN',
-      },
-      { isRequired: true, isInitialized: true }
-    );
-
-    // 2. 保存存储配置（可选）
-    if (setupData.storage) {
-      await setConfig('storage', setupData.storage, { isRequired: false, isInitialized: true });
-    }
-
-    // 3. 保存界面配置
-    await setConfig('ui', setupData.ui, { isRequired: false, isInitialized: true });
-
-    // 4. 生成并保存安全密钥
-    const securityKeysGenerated = await generateSecurityKeys();
-    if (!securityKeysGenerated) {
-      throw new Error('Failed to generate security keys');
-    }
-
-    // 6. 标记系统为已初始化
-    // 获取当前迁移版本
-    const migrationVersion = await getLatestMigrationVersion();
-
-    await setConfig(
-      'system',
-      {
-        isInitialized: true,
-        initializationVersion: '1.0.0',
-        lastMigrationVersion: migrationVersion,
-      },
-      { isRequired: true, isSystem: true, isInitialized: true }
-    );
-
-    // 7. 获取生成的密钥（用于响应）
-    const securityConfig = await getConfig('security');
-    const notificationConfig = await getConfig('notification');
-
-    const response: SetupResponse = {
-      success: true,
-      message: reusedExistingAdmin
-        ? 'System initialization completed using the existing admin user. The existing password was not changed.'
-        : 'System initialization completed successfully',
-      data: {
-        adminUser,
-        generatedKeys: {
-          jwtSecret: (securityConfig as any)?.jwtSecret || '',
-          jwtRefreshSecret: (securityConfig as any)?.jwtRefreshSecret || '',
-          sessionSecret: (securityConfig as any)?.sessionSecret || '',
-          vapidPublicKey: (notificationConfig as any)?.vapidPublicKey || '',
-          vapidPrivateKey: (notificationConfig as any)?.vapidPrivateKey || '',
-        },
-      },
-    };
-
-    return c.json(createResponse(response.data, response.message), 200);
-  } catch (error: any) {
-    console.error('System initialization failed:', error);
-    return c.json(createResponse(null, `System initialization failed: ${error.message}`), 500);
-  }
-});
 
 // 获取所有配置（管理员）
 adminRouter.get('/settings', authenticateJWT, requireAdmin, async (c: HonoContext) => {
@@ -765,123 +392,6 @@ adminRouter.post('/refresh-cache', authenticateJWT, requireAdmin, async (c: Hono
   }
 });
 
-// 获取数据看板统计（管理员）
-adminRouter.get('/stats/dashboard', authenticateJWT, requireAdmin, async (c: HonoContext) => {
-  try {
-    const stats = await getDashboardStats();
-    return c.json(createResponse(stats), 200);
-  } catch (error: any) {
-    console.error('Failed to get dashboard stats:', error);
-    return c.json(createResponse(null, 'Failed to get dashboard stats'), 500);
-  }
-});
-
-// 获取所有用户列表（管理员）
-adminRouter.get('/users', authenticateJWT, requireAdmin, async (c: HonoContext) => {
-  const page = c.req.query('page') || '1';
-  const limit = c.req.query('limit') || '10';
-  const role = c.req.query('role');
-  const search = c.req.query('search');
-  const sortField = c.req.query('sortField');
-  const sortOrder = c.req.query('sortOrder') as 'asc' | 'desc' | undefined;
-  const { users, total } = await listUsers({ page, limit, role, search, sortField, sortOrder });
-
-  return c.json(
-    createResponse({
-      users,
-      pagination: {
-        page: Number(page),
-        limit: Number(limit),
-        total,
-        pages: Math.ceil(total / Number(limit)),
-      },
-    }),
-    200
-  );
-});
-
-// 更新用户角色（超级管理员）
-adminRouter.put(
-  '/users/:userId/role',
-  authenticateJWT,
-  requireSuperAdmin,
-  async (c: HonoContext) => {
-    const userId = c.req.param('userId');
-    const body = await c.req.json();
-    const { role } = body as { role: UserRole };
-
-    if (!Object.values(UserRole).includes(role)) {
-      throw new Error('Invalid role');
-    }
-
-    const user = await updateUserRole(userId, role);
-
-    return c.json(createResponse(user, 'User role updated successfully'), 200);
-  }
-);
-
-// 获取用户详细信息（管理员）
-adminRouter.get('/users/:userId', authenticateJWT, requireAdmin, async (c: HonoContext) => {
-  const userId = c.req.param('userId');
-
-  const user = await getUserByIdForAdmin(userId);
-
-  if (!user) {
-    throw new Error('User not found');
-  }
-
-  return c.json(createResponse(user), 200);
-});
-
-// 删除用户（超级管理员）
-adminRouter.delete('/users/:userId', authenticateJWT, requireSuperAdmin, async (c: HonoContext) => {
-  const userId = c.req.param('userId');
-
-  await deleteUserById(userId);
-
-  return c.json(createResponse(null, 'User deleted successfully'), 200);
-});
-
-// 验证用户邮箱（管理员）
-adminRouter.put(
-  '/users/:userId/verify-email',
-  authenticateJWT,
-  requireAdmin,
-  async (c: HonoContext) => {
-    const userId = c.req.param('userId');
-
-    const user = await verifyUserEmail(userId);
-
-    if (!user) {
-      throw new Error('User not found');
-    }
-
-    return c.json(createResponse(user, 'User email verified successfully'), 200);
-  }
-);
-
-// 取消验证用户邮箱（管理员）
-adminRouter.put(
-  '/users/:userId/unverify-email',
-  authenticateJWT,
-  requireAdmin,
-  async (c: HonoContext) => {
-    const userId = c.req.param('userId');
-
-    const user = await unverifyUserEmail(userId);
-
-    if (!user) {
-      throw new Error('User not found');
-    }
-
-    return c.json(createResponse(user, 'User email unverified successfully'), 200);
-  }
-);
-
-// 获取角色统计信息（管理员）
-adminRouter.get('/roles/stats', authenticateJWT, requireAdmin, async (c: HonoContext) => {
-  const stats = await getRoleStats();
-  return c.json(createResponse(stats), 200);
-});
+adminRouter.route('/', adminUsersRouter);
 
 export default adminRouter;
